pyproject.toml
[project]
name = "remnasale"
version = "0.0.1"
description = "This is a Telegram bot for selling VPN subscriptions. It works with Remnawave"
authors = [{ name = "DanteFuaran" }]
readme = "README.md"
license = "MIT"
requires-python = ">=3.12,<3.13"
dependencies = [
    "aiogram~=3.25.0",
    "aiogram-dialog~=2.4.0",
    "alembic~=1.16.5",
    "asyncpg~=0.30.0",
    "dishka~=1.6.0",
    "fastapi>=0.120.2",
    "fluentogram~=1.2.1",
    "greenlet>=3.2.4",
    "loguru~=0.7.3",
    "msgspec~=0.19.0",
    "pydantic-settings~=2.11.0",
    "redis~=7.0.0",
    #"remnapy @ git+https://github.com/snoups/remnapy.git@development",
    "remnapy>=2.3.3",
    "taskiq[orjson]~=0.12.1",
    "taskiq-redis~=1.2.0",
    "uvicorn>=0.38.0",
    "cryptography>=46.0.3",
    "sqlalchemy[mypy]>=2.0.0",
    "qrcode[pil]>=8.2",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.20",
    "passlib[bcrypt]>=1.7.0",
    "python-jose[cryptography]>=3.3.0",
]

[project.urls]
homepage = "https://t.me/remnasale"
repository = "https://github.com/DanteFuaran/Remnasale"

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.14.2",
    "pytest>=8.4.2",
    "watchfiles>=1.1.1",
    "ftl-extract>=0.9.0",
    "types-cachetools",
]

[tool.ruff]
target-version = "py312"
line-length = 100
lint.select = [
    "C",
    "DTZ",
    "E",
    "F",
    "I",
    "ICN",
    "N",
    "PLC",
    "PLE",
    "Q",
    "T",
    "W",
    "YTT",
]
lint.ignore = ["N805"]
exclude = ["venv", ".venv", ".idea", "tests"]

[tool.mypy]
packages = ["src"]
plugins = ["sqlalchemy.ext.mypy.plugin", "pydantic.mypy"]
exclude = ["venv", ".venv", ".idea", "tests"]
warn_unused_configs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true
strict_equality = true
extra_checks = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = ["fluentogram.*", "remnawave.*", "qrcode.*", "remnapy.*"]
follow_untyped_imports = true

[[tool.mypy.overrides]]
module = ["app.bot.handlers.*"]
strict_optional = false
warn_return_any = false
disable_error_code = ["union-attr"]
